chore: move sign in related files to folders based on usage
