+++ packages/apps/human-app/frontend/src/api/fetcher.ts
@@ -3,8 +3,8 @@ import { ZodError, type ZodType, type ZodTypeDef } from 'zod';
 import type { ResponseError } from '@/shared/types/global.type';
 import { browserAuthProvider } from '@/shared/contexts/browser-auth-provider';
 import { env } from '@/shared/env';
-import { type SignInSuccessResponse } from '@/modules/worker/services/sign-in/types';
 import { normalizeBaseUrl } from '@/shared/helpers/url';
+import { type SignInSuccessResponse } from '@/shared/types';
 import { fetchTokenRefresh } from './fetch-refresh-token';
 
 const appendHeader = (
